+++ pkg/loadbalancer/redirectpolicy/testdata/node-local-dns.txtar
