+++ pkg/loadbalancer/redirectpolicy/testdata/service.txtar
