+++ pkg/loadbalancer/redirectpolicy/testdata/skiplb-addr.txtar
