+++ pkg/loadbalancer/redirectpolicy/watch.sh
